+++ autometrics/build.rs
@@ -6,16 +6,18 @@ pub fn main() {
     #[cfg(feature = "metrics")]
     println!("cargo:warning=The `metrics` feature is deprecated and will be removed in the next version. Please use `metrics-0_21` instead.");
     #[cfg(feature = "opentelemetry")]
-    println!("cargo:warning=The `opentelemetry` feature is deprecated and will be removed in the next version. Please use `opentelemetry-0_19` instead.");
+    println!("cargo:warning=The `opentelemetry` feature is deprecated and will be removed in the next version. Please use `opentelemetry-0_20` instead.");
     #[cfg(feature = "prometheus")]
     println!("cargo:warning=The `prometheus` feature is deprecated and will be removed in the next version. Please use `prometheus-0_13` instead.");
     #[cfg(feature = "prometheus-client")]
     println!("cargo:warning=The `prometheus-client` feature is deprecated and will be removed in the next version. Please use `prometheus-client-0_21` instead.");
+    #[cfg(feature = "exemplars-tracing-opentelemetry")]
+    println!("cargo:warning=The `exemplars-tracing-opentelemetry` feature is deprecated and will be removed in the next version. Please use `exemplars-tracing-opentelemetry-0_20` instead.");
 
     cfg_aliases! {
       // Backends
       metrics: { any(feature = "metrics", feature = "metrics-0_21") },
-      opentelemetry: { any(feature = "opentelemetry", feature = "opentelemetry-0_19") },
+      opentelemetry: { any(feature = "opentelemetry", feature = "opentelemetry-0_20") },
       prometheus: { any(feature = "prometheus", feature = "prometheus-0_13") },
       prometheus_client_feature: { any(feature = "prometheus-client", feature = "prometheus-client-0_21") },
       default_backend: { all(
@@ -30,7 +32,7 @@ pub fn main() {
       // Exemplars
       exemplars: { any(exemplars_tracing, exemplars_tracing_opentelemetry) },
       exemplars_tracing: { feature = "exemplars-tracing" },
-      exemplars_tracing_opentelemetry: { feature = "exemplars-tracing-opentelemetry" },
+      exemplars_tracing_opentelemetry: { any(feature = "exemplars-tracing-opentelemetry-0_20", feature = "exemplars-tracing-opentelemetry") },
 
       // Custom objectives
       custom_objective_percentile: { feature = "custom-objective-percentile" },