+++ autometrics/src/constants.rs
@@ -37,6 +37,3 @@ pub const COMMIT_KEY: &str = "commit";
 pub const BRANCH_KEY: &str = "branch";
 pub const SERVICE_NAME_KEY: &str = "service.name";
 pub const SERVICE_NAME_KEY_PROMETHEUS: &str = "service_name";
-
-// Units
-pub const HISTOGRAM_UNIT: &str = "seconds";