+++ CHANGELOG.md
@@ -38,13 +38,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `custom-objective-latency` feature can now be used with the `prometheus-exporter`, as well
   as with the `prometheus` and `prometheus-client` crates, because the histogram buckets can now
   be configured via the settings
+- Upgraded `opentelemetry` to v0.20
 
 ### Deprecated
 
 - `metrics` feature flag (replaced with `metrics-0_21`)
-- `opentelemetry` feature flag (replaced with `opentelemetry-0_19`)
+- `opentelemetry` feature flag (replaced with `opentelemetry-0_20`)
 - `prometheus` feature flag (replaced with `prometheus-0_13`)
 - `prometheus-client` feature flag (replaced with `prometheus-client-0_21`)
+- `exemplars-tracing-opentelemetry` feature flag (replaced with `exemplars-tracing-opentelemetry-0_20`)
 
 ### Removed